Update pre-install script. Will not re-install some packages which already installed
Former-commit-id: 942bc5262ce3227b53b2980a0a0331b0e8ca6b56
Former-commit-id: dc7dcf4a60ffd7a6bcc58f069f1f3177b5d13924
Former-commit-id: ed3a08568e9579f81d39db2ee1d2442984a1cc06

+++ vendors/pre-install.sh
@@ -2,36 +2,42 @@
 sudo apt-get install -y python-dev
 
 # Install pcre-8.37
-tar xvfz pcre-8.37.tar.gz
-cd pcre-8.37/
-./configure
-make
-sudo make install
-sudo rm -rf pcre-8.37/
-cd ../
-
+if [ ! -f /usr/local/lib/libpcre.a ]; then
+	tar xvfz pcre-8.37.tar.gz
+	cd pcre-8.37/
+	./configure
+	make
+	sudo make install
+	cd ../
+	sudo rm -rf pcre-8.37/
+fi
 # Install swig-3.0.5
-tar xvfz swig-3.0.5.tar.gz
-cd swig-3.0.5/
-./configure
-make
-sudo make install
-sudo rm -rf swig-3.0.5/
-cd ..
+if [ ! -d /usr/local/share/swig/ ]; then
+	tar xvfz swig-3.0.5.tar.gz
+	cd swig-3.0.5/
+	./configure
+	make
+	sudo make install
+	cd ..
+	sudo rm -rf swig-3.0.5/
+fi
 
 # Refresh ldconfig
 sudo ldconfig
 
 # Install libntoh
-cd libntoh/
-mkdir -p install
-cd src/
-mkdir -p build
-cd build/
-cmake .. -DCMAKE_INSTALL_PREFIX=../../install
-make
-sudo make install
-
+if [ ! -f /usr/local/lib/libntoh.a ]; then
+	cd libntoh/
+	mkdir -p install
+	cd src/
+	mkdir -p build
+	cd build/
+	cmake .. 
+	make
+	sudo make install
+	cd ..
+	sudo rm -rf build/
+	cd ../../
+fi
 # Back to vendors/
-cd ../../../
 sudo ldconfig
